refactor: move sudo builder to its own directory
